# docs/openapi.yaml
openapi: 3.0.3
info:
  title: qf Control Plane API
  description: |
    REST API for the qf host-based firewall control plane.

    **Authentication**: Most endpoints require either a JWT cookie (`qf_token`, set by `POST /auth/login`)
    or an `Authorization: Bearer <token>` header (JWT or opaque API token).

    **Tenant**: Protected endpoints require the `X-Tenant-ID` header (UUID of the tenant).
  version: "0.5.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: qf_token
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant UUID
      schema:
        type: string
        format: uuid
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Token:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [single_host, bulk]
        target_host_id:
          type: string
          format: uuid
          nullable: true
        label_template:
          type: object
          additionalProperties:
            type: string
        max_uses:
          type: integer
        uses_count:
          type: integer
        expires_at:
          type: string
          format: date-time
        token:
          type: string
          description: Plain token value — only present in create response
    CreateTokenRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [single_host, bulk]
        target_host_id:
          type: string
          format: uuid
          description: Required when type=single_host
        label_template:
          type: object
          additionalProperties:
            type: string
        ttl_seconds:
          type: integer
          default: 3600
        max_uses:
          type: integer
          default: 1

    # ── Default Policy ─────────────────────────────────────────────────────
...
    APIToken:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
          enum: [admin, operator, auditor]
        created_by:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
          nullable: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        token:
          type: string
          description: Plain token — only present in create response
    CreateAPITokenRequest:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, operator, auditor]
        expires_at:
          type: string
          format: date-time
          nullable: true

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using refresh cookie
      security: []
      responses:
        "200":
          description: New access token cookie set
        "401":
          description: Invalid or expired refresh token
...
  /tokens:
    get:
      tags: [Tokens]
      summary: List enrollment tokens
      parameters:
        - $ref: '#/components/parameters/TenantID'
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Token'
    post:
      tags: [Tokens]
      summary: Create an enrollment token
      parameters:
        - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTokenRequest'
      responses:
        "201":
          description: Token created; plain token in `token` field (shown once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
  /tokens/{id}:
    delete:
      tags: [Tokens]
      summary: Revoke an enrollment token
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - $ref: '#/components/parameters/ResourceID'
      responses:
        "204":
          description: Revoked

  # ── Default Policy ────────────────────────────────────────────────────────
...
  /api-tokens:
    get:
      tags: [APITokens]
      summary: List API tokens
      responses:
        "200":
          description: API tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIToken'
    post:
      tags: [APITokens]
      summary: Create an API token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPITokenRequest'
      responses:
        "201":
          description: Token created; plain token in `token` field (shown once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIToken'
  /api-tokens/{id}:
    delete:
      tags: [APITokens]
      summary: Revoke an API token
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        "204":
          description: Revoked

  # ── OpenAPI + Docs ────────────────────────────────────────────────────────
...